refactor: update dependency and edited import file path. added prisma schema.
